+++ app/src/main/java/com/example/memorycat/Repository/Repository.kt
@@ -9,9 +9,11 @@ class Repository {
     private val _userDB = _firestore.collection("userDB").document(uid!!)
     private val _accureDB = _firestore.collection("accurequizDB").document(uid!!)
     private val _recentDB = _firestore.collection("recentquizDB").document(uid!!)
+    private val _bookmarkDB = _firestore.collection("recentbookmarkDB").document(uid!!)
 
     val firestore get() = _firestore
     val userDB get() = _userDB
     val accureDB get() = _accureDB
     val recentDB get() = _recentDB
+    val bookmarkDB get() = _bookmarkDB
 }
